agent/requirements.txt
scapy
pandas
numpy
scikit-learn
joblib
requests
cryptography
websockets
flask
pywinrm
paramiko
